first version of a SST enabled B2B application.
The incoming INVITE establishind a call is passed in 
signaling only B2B mode to the B leg, which tries to 
send it to the request URI.

SIP Session Timers are enabled on both legs. When the 
timer expires, an empty INVITE is sent, and the resulting 
SDP offer from body of the 200 is relayed into the other 
leg, where it is sent out as INVITE with the offer. The 
answer from B leg is relayed into A leg and sent as body
in ACK message.

SST expiration is configurable in confi file. 



git-svn-id: http://svn.berlios.de/svnroot/repos/sems/trunk@1821 8eb893ce-cfd4-0310-b710-fb5ebe64c474