b/f: send autheticated BYE with To-tag
closes https://bugtracker.iptel.org/view.php?id=44
Summary:                    authenticated BYE does not work (to-tag missing)
Description:
When authentication is requested by a proxy for a BYE request sent by SEMS,
the second BYE request (with authentication info) does not contain a
to-tag. This causes the proxy to reject the request with a 481 reply.

Thanks to <samusenko at msm dot ru> for reporting this bug.