b/f: B2BUA: don't refresh for establishing transaction
fixes click2dial application.

In the B2BUA, usually a refresh is done if the established session
description has changed on the other leg (e.g. initiated by SST
refresh from the middle). This fix prevents that refresh for the
initial transaction, where the re-INVITE to caller is done explicitely.