b/f: transcoder: reusing payload IDs of transcodable codecs from SDP offer
When adding transcoder codecs into SDP answer try to use their IDs from SDP
offer. Though it should work with newly assigned IDs it doesn't.

This code is just approximation - offer/answer is not detected properly and
payload IDs are remembered for longer than just offer/answer exchange but it
should work. It doesn't matter if we assign such or another payload IDs as long
as they are not already in use.