dtmf application level relay and transcoding
In case the transcoder is turned on, RTP DTMF are relayed at application level so that they are understood as coming from the same source from the receiver. When the DTMF transcoding is turned on, in band DTMF received on one leg are re-sent as RTP DTMFs on the other leg. However, the audio signals are not filtered. Beware that the detection lag can cause a receiver capable to understand both signals to capture those DTMFs twice, as there is a big lag between the inband and RTP DTMFs.