b/f: fixes a possible memory leak in the event dispatcher.
Until this fix, it was possible to add an event queue with local tag, called and via-branch-tag, but omitting those parameters while deleting the event queue, thus leaving the lookup entries in the id_lookup map. This is no longer possible as the id_lookup map is now managed automatically. The additional id used as an index in id_lookup is saved together with the main queue entry in the queue map.