SBC: CC module initialization with values from SBC profile instead of initial INVITE request
Without the INVITE request module parameters can not be evaluated during
initialization but the values from SBC profile are already evaluated and should
be used instead.

The reason for changing this is that we can add new call legs during the call
life when doing blind call transfer and we need not to remember the initial
INVITE to be able to do that.