new interface config
Allows for signaling-only or media-only interfaces.

The parameter "additional_interfaces" has been renamed to
"interfaces". When not present, SEMS will look for IP config
parameters without any suffixes (sip_ip, media_ip, etc...).

If the "media_ip_[interface name]" is set, the interface is considered
to be a media interface.  If the "sip_ip_[interface name]" parameter
is defined, the interface is considered to be a media interface.  If
both parameters are defined for the same interface, it is a signaling
and media interface.