sbc: quick hack to support early media and 100rel again
Correct support for early media and 100rel together with forking would require
creating multiple A legs towards the caller but this would be pretty tricky
with current design.

So simplified approach is chosen - first of the B legs is used and its media
and replies are relayed back to the caller. Upon error reply from such B leg
another may be chosen (including SDP). This is against SIP (see 13.2.1 in RFC
3261) but suprisingly seems to work.