sip: destination blacklist
When a destination cannot be reached (timeout+grace period), the IP/port tuple is blacklisted for a certain period of time (default_bl_ttl). if all destinations for a message are blacklisted, a 500 error reply is generated internally.

In case a downstream proxy or UA returns 503, this proxy or UA also gets blacklisted.