sbc:rtprelay_dtmf_filtering and rtprelay_dtmf_detection options
With the option rtprelay_dtmf_filtering=yes the SBC filters out RTP DTMF
(RFC2833 / RFC4733) packets in relayed streams.

If rtprelay_dtmf_detection=yes is set, DTMF from RTP packets is detected
and can be used to control applications, e.g. special call flows, implemented with
the extended call control API. Note that the call needs to be added to
the media processor in order for DTMF events to be processed.

Internally, if the flag force_dtmf_receiving is set for the RTP stream,
also in relay mode the DTMF packets are sent to the DTMF detection queue.

The ext cc api is extended with the onDtmf function.