core: b/f: fixed possible crash in AmEventQueueProcessor
In case an event is posted to an already finalised queue, the queue processor would cause a seg-fault because the queue pointer would still be in its processing queue. This has been fixed by using ref-counting.

If the queue processor is used, the queue's ref-count should be increased before binding the queue to the processor. Also, the ref-count should be decreased at some place during or after finalize(), so that the queue is disposed correctly.