sip: b/f: find the correct transaction to cancel
In case the request has been sent to multiple destinations for failover purposes, new transactions might be created. To cancel such transactions, it necessary to find the last one by looking at the dialog-id and cseq, as only a reference to the original transaction has been saved in the dialog state.