core: fix multipart delimiter parser
This fixes a bug happening when a body-part of a multipart body ends with
CRLF: in this case SEMS considers the body-part ending CRLF(s) as being
part of the delimiter.

A delimiter starts with only one CRLF, so we have to reset the pointer
to the delimiter begin during a transition from B_LF to B_CR state.